backend/requirements.txt
Django==5.2.1
asgiref==3.8.1
djangorestframework==3.16.0
pip==23.2.1
sqlparse==0.5.3
tzdata==2025.2
djoser==2.3.1
pillow==11.2.1
django-cors-headers==4.7.0
python-dotenv==1.1.0
